[FLINK-33341][state] Refactoring: consolidate equals/hashCode/toString for incremental state handle classes.
